+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/di/AppModule.kt
@@ -4,6 +4,7 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import jp.co.yumemi.android.code_check.api.Api
 import jp.co.yumemi.android.code_check.repository.SearchRepository
 import jp.co.yumemi.android.code_check.repository.SearchRepositoryImpl
 import javax.inject.Singleton
@@ -13,5 +14,9 @@ import javax.inject.Singleton
 object AppModule {
     @Provides
     @Singleton
-    fun provideSearchRepository(): SearchRepository = SearchRepositoryImpl()
+    fun provideSearchRepository(): SearchRepository = SearchRepositoryImpl(api = Api())
+
+    @Provides
+    @Singleton
+    fun provideApi(): Api = Api()
 }
