+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/repository/SearchRepositoryImpl.kt
@@ -1,31 +1,20 @@
 package jp.co.yumemi.android.code_check.repository
 
-import io.ktor.client.*
-import io.ktor.client.call.*
-import io.ktor.client.engine.android.*
-import io.ktor.client.request.*
-import io.ktor.client.statement.*
+import jp.co.yumemi.android.code_check.api.Api
 import jp.co.yumemi.android.code_check.entity.RepositoryInfo
 import jp.co.yumemi.android.code_check.entity.Resource
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import org.json.JSONObject
 
-class SearchRepositoryImpl : SearchRepository {
+class SearchRepositoryImpl(private val api: Api) : SearchRepository {
 
     private val _state = MutableStateFlow<Resource<List<RepositoryInfo>>>(Resource.Empty)
     override val state: StateFlow<Resource<List<RepositoryInfo>>> = _state
 
-    override suspend fun fetchRepository(inputText: String) {
+    override suspend fun fetchRepository(searchText: String) {
         try {
-            val client = HttpClient(Android)
-            val response: HttpResponse =
-                client.get("https://api.github.com/search/repositories") {
-                    header("Accept", "application/vnd.github.v3+json")
-                    parameter("q", inputText)
-                }
-            val jsonBody = JSONObject(response.receive<String>())
-            val jsonItems = jsonBody.optJSONArray("items")
+            val jsonItems = api.fetchRepository(searchText)
             val items = mutableListOf<RepositoryInfo>()
 
             jsonItems?.let {
